requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
colorama==0.4.0
cryptography==2.3.1
future==0.17.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycparser==2.19
pylint==2.1.1
pyOpenSSL==18.0.0
PySocks==1.6.8
python-telegram-bot==11.1.0
requests==2.19.1
six==1.11.0
ujson==1.35
urllib3==1.23
win-inet-pton==1.0.1
wincertstore==0.2
wrapt==1.10.11
